remove(remix): Drop Remix v1 support
