+++ src/remix/codemods/root.ts
@@ -2,18 +2,78 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 /* eslint-disable @typescript-eslint/no-unsafe-argument */
-
 import * as recast from 'recast';
 import * as path from 'path';
 
 import type { ExportNamedDeclaration, Program } from '@babel/types';
 
-// @ts-expect-error - magicast is ESM and TS complains about that. It works though
-import { loadFile, writeFile } from 'magicast';
+import {
+  builders,
+  generateCode,
+  loadFile,
+  ProxifiedModule,
+  writeFile,
+  // @ts-expect-error - magicast is ESM and TS complains about that. It works though
+} from 'magicast';
 
 import { ERROR_BOUNDARY_TEMPLATE_V2 } from '../templates';
 import { hasSentryContent } from '../utils';
-import { wrapAppWithSentry } from './root-common';
+import chalk from 'chalk';
+
+// @ts-expect-error - clack is ESM and TS complains about that. It works though
+import * as clack from '@clack/prompts';
+
+export function wrapAppWithSentry(
+  rootRouteAst: ProxifiedModule,
+  rootFileName: string,
+) {
+  rootRouteAst.imports.$add({
+    from: '@sentry/remix',
+    imported: 'withSentry',
+    local: 'withSentry',
+  });
+
+  recast.visit(rootRouteAst.$ast, {
+    visitExportDefaultDeclaration(path) {
+      if (path.value.declaration.type === 'FunctionDeclaration') {
+        // Move the function declaration just before the default export
+        path.insertBefore(path.value.declaration);
+
+        // Get the name of the function to be wrapped
+        const functionName: string = path.value.declaration.id.name as string;
+
+        // Create the wrapped function call
+        const functionCall = recast.types.builders.callExpression(
+          recast.types.builders.identifier('withSentry'),
+          [recast.types.builders.identifier(functionName)],
+        );
+
+        // Replace the default export with the wrapped function call
+        path.value.declaration = functionCall;
+      } else if (path.value.declaration.type === 'Identifier') {
+        const rootRouteExport = rootRouteAst.exports.default;
+
+        const expressionToWrap = generateCode(rootRouteExport.$ast).code;
+
+        rootRouteAst.exports.default = builders.raw(
+          `withSentry(${expressionToWrap})`,
+        );
+      } else {
+        clack.log.warn(
+          chalk.yellow(
+            `Couldn't instrument ${chalk.bold(
+              rootFileName,
+            )} automatically. Wrap your default export with: ${chalk.dim(
+              'withSentry()',
+            )}\n`,
+          ),
+        );
+      }
+
+      this.traverse(path);
+    },
+  });
+}
 
 export async function instrumentRootRouteV2(
   rootFileName: string,
